// src/types/pnl.ts
export interface MonthlyPnL {
  month: string
  newSubscriptions: number
  expansion: number
  renewals: number
  professionalServices: number
  refundsAndCredits: number
  totalRevenue: number
  hostingAndInfrastructure: number
  customerSuccess: number
  totalCostOfRevenue: number
  grossProfit: number
  grossMarginPct: number
  salesAndMarketing: number
  researchAndDevelopment: number
  generalAndAdministrative: number
  totalOperatingExpenses: number
  operatingIncome: number
  operatingMarginPct: number
}